Fix `findOneByKey` repository method
Converting the new uuid system in the previous commit, means that the
uuids are stored as binary data. When using query builders, this is not
a problem, symfony does the conversion automatically. But when we use a
native query (like we do in this method), we have to convert the ids to
binary data ourselves, before sending them to the database as query
parameter.